Remove gustiness from U10 and add U10WITHGUSTS.
The E3SM wind at the 10 meter reference level (an output from the
atmosphere, land, and sea ice components) has previously included the
effects of gustiness, but this is not correct for comparison to most
other data sets. This commit therefore removes the effect of gustiness
from these outputs. See #5836.

Note that MPAS-Ocean does not output such a wind itself, only the
mean-squared wind, which is used solely to determine flux of tracers
and should in fact include gustiness. Therefore, MPAS-Ocean has not
been changed.

This change applies to both atmosphere-calculated gustiness, and the
land-calculated gustiness used in cases where ELM is active but EAM is
not. (E.g. in ELM, both `ugust` from the atmosphere and `wc` from ELM
are removed from ELM's u10 calculation.) The removal of the gustiness
is accomplished by changes that effectively multply u10 by a factor of
`ubot / sqrt(ubot^2 + ugust^2 + wc^2)`
where `ubot` is the resolved wind speed in the atmosphere's lowest
level. Similar expressions (without the `wc` term) are used in the
ocean coupling code and MPAS-Seaice.

For those cases where a gustiness-enhanced output may be useful, for
instance where it is desirable to know which wind was used to
calculate surface energy fluxes, or for comparison to previous model
versions. This commit therefore adds the diagnostic output
U10WITHGUSTS to EAM. This is defined identically to to the old
definition of U10.

[non-BFB], but only the U10 outputs of EAM and ELM, and
atmosReferenceSpeed10m in MPAS-Seaice, should be affected, and only in
cases where EAM or ELM.