Add support for atmosphere-supplied gustiness to MPAS-Seaice.
When ATM_SUPPLIES_GUSTINESS=TRUE, the atmosphere model provides a "ugust" field
that should be used to enhance mean winds for certain surface flux calculations.
Support for this feature was added to CICE in the past; this commit ports those
changes to MPAS-Seaice. It also is part of the fix for #5834, which requires
all components that calculate fluxes to the atmosphere to respect
ATM_SUPPLIES_GUSTINESS.

[CC] when ATM_SUPPLIES_GUSTINESS=TRUE (i.e. for EAM runs), otherwise
bit-for-bit.