Push gustiness calculations down into surface models.
Previously, the `ATM_SUPPLIES_GUSTINESS` XML variable had two effects:
 1) It caused EAM to export the variable `ugust` instead of applying
    it to the exported wind variables.
 2) It caused flux calculations for select components (ELM, CICE, and
    the ocean model) to account for the effects of gustiness. For
    these components, issue #5834 would be fixed, while gustiness for
    all other components would be turned off.

There are two problems that result from this. First, it combines a
non-climate-changing effect (changing the location where `ugust` is
applied) with a climate-changing effect (a partial fix for #5834).
Second, because this option has not been maintained, the option has
the effect of turning off gustiness effects entirely for WW3 and
MPAS-SeaIce; this behavior was never intended.

Nonetheless, addressing issue #5834 will ultimately require the
atmosphere to supply a gustiness fields to other components, and so
the core functionality of `ATM_SUPPLIES_GUSTINESS` is still needed.

This PR removes the climate-changing effects of
`ATM_SUPPLIES_GUSTINESS` Additionally, `ATM_SUPPLIES_GUSTINESS` is now
enabled for all EAM compsets (but not SCREAM compsets).

This PR should be considered to be a refactor laying the groundwork
for future fixes to the gustiness parameterization. For compsets where
EAM is not running, or configurations where all components run using
the same horizontal grid and time step size (mainly F compsets), this
commit is bit-for-bit. For all other configurations, this will be
non-BFB but not climate-changing. The change in answers comes from
changing order of operations, namely regridding information in the
coupler before applying `ugust`, rather than afterward.

[non-BFB]
[NML]